facebook has reached (eol) with the nuke communities!
Facebook has proven useless to the nuke communities and is ever changing and constantly breaking websites. When facebook did away with their like system, 1 million programmers did away with the use of Facebook for any purpose in almost all the developer communities.

Facebook is garbage! Just one of the ways Zucker Fuck keeps getting richer. We will not support this winey punk...